Rework Express Integration
Having the primary class be an event emitter for request acceptance / rejection just didn't feel very clean when we already had to use a direct delegate method to get the API key.
Migrate all of the express-specific code out of the main Server class and into a subclass, also add a new export function that can create all the middleware we need in one execution.